more progress on the profile page
